Move CryptString to MiqPassword::Key
