+++ lib/gems/pending/metadata/VMMount/VMPlatformMountLinux.rb
@@ -1,4 +1,4 @@
-require 'encryption/CryptString'
+require 'util/miq-password'
 
 module VMPlatformMountLinux
   VCBNAME     = '/usr/sbin/vcbVmName'
@@ -32,7 +32,7 @@ module VMPlatformMountLinux
     # TODO: passing the user name and password on the command line is BAD.
     #       Change this to use web-services.
     #
-    @cs = CryptString.new
+    @cs = MiqPassword::Key.new
     cmd = "#{VCBNAME} -h #{@ems['host']} -u #{@ems['user']} -p #{@cs.decrypt(@ems['password'])} -s \"Name:#{@vmName}\" | grep \"moref:\""
     #
     # XXX Don't forget to remove this. We're logging the username and password, not good.
