+++ lib/gems/pending/util/miq-password.rb
@@ -1,5 +1,5 @@
 require 'ezcrypto'
-require 'encryption/CryptString'
+require 'openssl'
 require 'base64'
 require 'yaml'
 
@@ -188,12 +188,45 @@ EOS
       EzCrypto::Key.load(filename)
     else
       params = YAML.load_file(filename)
-      CryptString.new(nil, params[:algorithm], params[:key], params[:iv])
+      Key.new(nil, params[:algorithm], params[:key], params[:iv])
     end
   end
 
   def self.extract_erb_encrypted_value(value)
     return $1 if value =~ /\A<%= (?:MiqPassword|DB_PASSWORD)\.decrypt\(['"]([^'"]+)['"]\) %>\Z/
+  end
+
+  class Key
+    def initialize(_str = nil, enc_alg = nil, key = nil, iv = nil)
+      @enc_alg = enc_alg
+      @key     = key
+      @iv      = iv
+    end
+
+    def encrypt64(str)
+      cip = OpenSSL::Cipher::Cipher.new(@enc_alg)
+      cip.encrypt
+      cip.key = @key
+      cip.iv  = @iv
+
+      es = cip.update(str)
+      es << cip.final
+      [es].pack('m')
+    end
+    alias_method :encrypt, :encrypt64
+
+    def decrypt64(str)
+      cip = OpenSSL::Cipher::Cipher.new(@enc_alg)
+      cip.decrypt
+      cip.key = @key
+      cip.iv  = @iv
+
+      rs = cip.update(str.unpack('m').join)
+      rs << cip.final
+
+      rs
+    end
+    alias_method :decrypt, :decrypt64
   end
 end
